feat: migrate Dropdown component from jsx to tsx (#897)
